Mark a failing test as work_in_progress: test_c6h604 in pathfinderTest
The InChI in this test seems to be malformed. The old version of OpenBabel
could read it, but the resulting molecule has an invalid AdjacencyList.

Neither the new OpenBabel implementation nor RDKit can read it.

It was introduced in https://github.com/ReactionMechanismGenerator/RMG-Py/commit/76d7cbabf6f0e2e5062d2cf16cbece92d01f70f8
If anyone can figure out what it's meant to be and replace it, we can fix this test.