Use conda-forge's cclib not the RMG channel one.
The RMG one is out of date and uses the old OpenBabel import method.
The patch for the imports is at 
https://github.com/cclib/cclib/commit/3f73c94d1cb5c62fe5ba15af46e204fd2abab9da 
if someone wants to apply it to the "RMG" version of cclib and push new binaries to conda, 
but there may be other patches needed too, and it'd be preferable to get back on track 
and use the official cclib.

This is an attempt to do just that. Hopefully it works.
(Are there other differences in "RMG"'s version of cclib?)