Remove Surface_Recombination family from documentation.
This family was deleted in https://github.com/ReactionMechanismGenerator/RMG-database/pull/457
so should not be listed in the documentation.

I think some other families have been recently added without appearing here.
I'm not sure whether this table can be automatically generated, and/or when it should be updated. 
Begs a bigger question about documentation reflecting the latest development version or the previous release. 
But I'll leave that for a separate issue.